Improve error reporting with `datacube product add`
It's not uncommon to include `---` at the top or bottom of yaml files,
allowing multiple documents to be stored in the same file. The code for
adding products would see this and attempt to validate an *empty* doc,
which would fail with a big error message, and an incomplete stack
trace.

This now politely handles the case of empty documents, and if we do
find an invalid document, return the full stack trace in case of an
error.

I think we should go further still in returning useful error messages
when documents don't match the relevant schema, but the best way to do
that is to refactor the @schema_validated decorator.