Added input validation for analytics client.
The input should only contain `data` or `user_tasks`, otherwise a ValueError is raised.
Added the corresponding test.
Changed the currents tests to use function_params to send the shapefile, which will copy the required files, instead of relying on them being on the same filesystem for both client and worker.