Moved file transfer functions to its own module.
This is to untangle the execution engine and client, and allow single point of change for compression and serialisation.
Also simplified returned values to always be URIs and created a degraded mode where auxilliary files are returned via redis, by saving them as a compressed tar archive. This should be used for testing only.