Consolidation of sub-result geo coordinated in AE/EE Lazy array.
Sub-results return xarray descriptors for their own cell, part of the overall result cube. When geo coordinates are used, they are stitched together as sub-results come in, until the LazyArray has a complete set of coordinates along each of the dimensions. The utility method to return a dict-like descriptor of a DataArray was moved to `engine_common/xarray_utils.py`.