Continued refactoring parameter passing through S3.
Parameters now passed into and out of the execution engine using a uniform S3 packer, which works for both real and file S3. All exchanges are then reduced to a single URL of the payload file in S3. Also implemented most of the cleaning mechanisms, but some parts still pending.